+++ Tests/Program.cs
@@ -5,18 +5,15 @@ using System;
 namespace PythonShell.Tests {
     internal class Program {
         static async System.Threading.Tasks.Task Main(String[] args) {
-            var shell = new Shell {
-                Config = new ShellConfig {
-                    DefaultWorkingDirectory = "Scripts",
-                    PythonRequires = new String[] { "PySide6" }
-                },
-                // CreateDefaultEnv = true
-            };
+            var shell = new PythonShell(new PythonShellConfig {});
             await shell.Initialize();
 
-            await shell.RunFile("Scripts/test.py", useInstance: true, instanceName: "testInstance1", listener: new Listener.ShellListener {
-                OnComplete = () => shell.Clear()
-            });
+            var instance = Shell.ShellManager.GetInstance("default");
+            // instance.InstallRequires(new String[] { "PySide6" });
+            instance.InstallRequiresFromFile("Scripts/requirements.txt");
+            await instance.RunFile("Scripts/test.py", workingDirectory: "Scripts");
+
+            Console.WriteLine("finished");
         }
     }
 }