+++ CHANGELOG.md
@@ -6,3 +6,10 @@
 ## 0.0.2
 - Fixing bugs that occur in windows.
 - Add logs.
+
+<br />
+
+## 0.0.3
+- change directionality of library.
+- add 'ShellManager', 'ShellInstance'.
+- update 'README.md'